# advertising_app/infrastructure/entry-points/api-rest/src/main/java/co/com/agency/api/dealer/model/DealerRequest.java
package co.com.agency.api.dealer.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.springframework.validation.annotation.Validated;

import javax.validation.constraints.Size;

@Validated
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class DealerRequest {

    @JsonProperty("name")
    @Schema(example = "Dealer name test")
    @Size(max = 255)
    private String name;
}