# advertising_app/infrastructure/entry-points/api-rest/src/main/java/co/com/agency/api/tierLimit/model/TierLimitRequest.java
package co.com.agency.api.tierLimit.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.springframework.validation.annotation.Validated;

import javax.validation.constraints.Size;

@Validated
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class TierLimitRequest {

    @JsonProperty("tier")
    @Schema(example = "tier limit")
    int tier;
}